+++ src/Thumbnails/index.types.ts
@@ -22,11 +22,4 @@ export type SpinoramaThumbnailsProps = BoxProps & {
 	 * @default false
 	 */
 	reverse?: boolean
-
-	/**
-	 * Determines the currently selected slide
-	 * A number representing the index of the selected slide
-	 * If not provided, the first slide will be selected
-	 */
-	selected?: number
 }