+++ src/Wrapper/index.tsx
@@ -3,13 +3,14 @@ import { Box } from '@mui/material'
 
 // Types
 import type { SpinoramaWrapperProps } from './index.types'
+import type { SpinoramaItemProps } from '../Item/index.types'
 
 // Components
 const SpinoramaItem = React.lazy(() => import('../Item'))
 
 const SpinoramaWrapper: React.FC<SpinoramaWrapperProps> = (props: SpinoramaWrapperProps) => {
 	// Props
-	const { justify, reverse, selected, className, children } = props
+	const { justify, reverse, className, children } = props
 
 	// Varaibles
 	const justifyContent = justify === 'end' ? 'flex-end' : 'flex-start'
@@ -19,17 +20,9 @@ const SpinoramaWrapper: React.FC<SpinoramaWrapperProps> = (props: SpinoramaWrapp
 			<Box className="spinorama-items" flexDirection={reverse ? 'row-reverse' : 'row'} justifyContent={justifyContent} display="flex" minWidth={1} height="100%">
 				{React.Children.map(children, (child, index) => {
 					if (React.isValidElement(child)) {
-						// Check is item
-						const isItem = child.type.toString().indexOf('spinorama-item') > -1
-
 						// Clone element
-						return isItem ? (
-							React.cloneElement(child as React.ReactElement<any>, { selected: selected === index })
-						) : (
-							<SpinoramaItem key={index} selected={selected === index}>
-								{child}
-							</SpinoramaItem>
-						)
+						if (child.type.toString().indexOf('spinorama-item') > -1) return React.cloneElement(child as React.ReactElement<SpinoramaItemProps>, {})
+						else return <SpinoramaItem key={index}>{child}</SpinoramaItem>
 					} else return child
 				})}
 			</Box>